+++ tests/functional/RegistrationCest.php
@@ -1,6 +1,7 @@
 <?php
 
 
+use dektrium\user\service\RegistrationService;
 use dektrium\user\service\UserConfirmation;
 use dektrium\user\models\Token;
 use dektrium\user\models\User;
@@ -19,11 +20,12 @@ class RegistrationCest
 
     public function _after(FunctionalTester $I)
     {
-        \Yii::$container->set(Module::className(), [
-            'enableGeneratingPassword' => false,
-        ]);
         Yii::$container->set(UserConfirmation::className(), [
             'isEnabled' => true,
+        ]);
+        Yii::$container->set(RegistrationService::className(), [
+            'isEnabled' => true,
+            'isPasswordGenerated' => false,
         ]);
     }
 
@@ -33,9 +35,6 @@ class RegistrationCest
      */
     public function testRegistration(FunctionalTester $I)
     {
-        \Yii::$container->set(Module::className(), [
-            'enableGeneratingPassword' => false,
-        ]);
         Yii::$container->set(UserConfirmation::className(), [
             'isEnabled' => false,
         ]);
@@ -92,8 +91,8 @@ class RegistrationCest
      */
     public function testRegistrationWithoutPassword(FunctionalTester $I)
     {
-        \Yii::$container->set(Module::className(), [
-            'enableGeneratingPassword' => true,
+        Yii::$container->set(RegistrationService::className(), [
+            'isPasswordGenerated' => true,
         ]);
         Yii::$container->set(UserConfirmation::className(), [
             'isEnabled' => false,